Add per-route metadata for reading pages
